Added hide terminal on child exit
I liked the behaviour in 0.14 where I could exit a child (ie: hit Ctrl+D to logout of an stjerm) and the terminal would hide, but would re-create a fresh shell which would sit hidden waiting for me to use it again in the future. This commit adds that behaviour back.